create-job-flow tool is now a deprecated alias for create-cluster
