+++ mrjob/cmd.py
@@ -18,8 +18,11 @@ from __future__ import print_function
 from sys import argv
 from sys import stderr
 
+# map from command name to function to call
 commands = {}
 
+# map from command name to description for help
+# (deprecated commands won't be included here)
 descriptions = {}
 
 usage = """usage: mrjob {subcommand|--help}"
@@ -31,21 +34,26 @@ def error(msg=None):
     if msg:
         print(msg, file=stderr)
 
-    longest_command = max(len(name) for name in commands)
+    longest_name = max(len(name) for name in descriptions)
 
     def subcommand_line(name):
-        spaces = ' ' * (longest_command - len(name))
+        spaces = ' ' * (longest_name - len(name))
         return '  %s: %s%s' % (
             name, spaces, descriptions[name])
     print(usage, file=stderr)
     print('\n'.join(
-        subcommand_line(name) for name in sorted(commands)), file=stderr)
+        subcommand_line(name) for name in sorted(descriptions)), file=stderr)
 
 
-def command(name, description):
+def command(name, description=None):
+    """Decorate a function used to call a command.
+
+    If you don't set *description*, it won't be included in help
+    (useful for deprecated commands)."""
     def decorator(f):
         commands[name] = f
-        descriptions[name] = description
+        if description:
+            descriptions[name] = description
         return f
     return decorator
 
@@ -72,15 +80,22 @@ def audit_usage(args):
     main(args)
 
 
-@command('boss', 'Run a command on every node of a cluster.')
-def mrboss(args):
-    from mrjob.tools.emr.mrboss import main
+@command('create-cluster', 'Create a persistent EMR cluster')
+def create_cluster(args):
+    from mrjob.tools.emr.create_cluster import main
     main(args)
 
 
-@command('create-job-flow', 'Create an EMR job flow')
+# deprecated alias for create-cluster, delete in v0.6.0
+@command('create-job-flow')
 def create_jf(args):
     from mrjob.tools.emr.create_job_flow import main
+    main(args)
+
+
+@command('boss', 'Run a command on every node of a cluster.')
+def mrboss(args):
+    from mrjob.tools.emr.mrboss import main
     main(args)
